+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 ## 0.1.15-SNAPSHOT
 
 - better pool locking
+- transduce reducer
+
+TODO:
+- update fold
+- refactor transduce
+
 
 ## 0.1.14